test_simple_multiview.py
"""
Simple test for multiview generation using a lighter approach
Since Qwen models are very large (~10GB), let's test with a simpler method first
"""

import torch
from PIL import Image
import numpy as np
from pathlib import Path
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def create_simple_multiview_demo(image_path: str, output_dir: str = "./output"):
    """
    Create a simple demo multiview output by rotating/transforming the input image.
    This is a placeholder until we can download the full Qwen models.
    """
    logger.info("=" * 60)
    logger.info("Simple Multiview Demo (Placeholder)")
    logger.info("=" * 60)
    
    # Create output directory
    output_path = Path(output_dir)
    multiview_path = output_path / "multiview"
    multiview_path.mkdir(parents=True, exist_ok=True)
    
    # Load image
    logger.info(f"Loading image: {image_path}")
    img = Image.open(image_path).convert("RGB")
    logger.info(f"Image size: {img.size}")
    
    # Define 8 camera positions (standard_8view)
    camera_positions = [
        (0, 0, 1.0, "front view"),
        (45, 0, 1.0, "front-right quarter view"),
        (90, 0, 1.0, "right side view"),
        (135, 0, 1.0, "back-right quarter view"),
        (180, 0, 1.0, "back view"),
        (225, 0, 1.0, "back-left quarter view"),
        (270, 0, 1.0, "left side view"),
        (315, 0, 1.0, "front-left quarter view"),
    ]
    
    logger.info(f"\nGenerating {len(camera_positions)} placeholder views...")
    
    for idx, (azimuth, elevation, distance, description) in enumerate(camera_positions):
        # For now, just create simple transformations as placeholders
        # In reality, Qwen Multi-Angles LoRA would generate these
        
        # Apply simple rotation/transformation as demo
        if azimuth == 0:
            view_img = img  # Front view - original
        elif azimuth == 180:
            view_img = img.transpose(Image.FLIP_LEFT_RIGHT)  # Back - flip
        elif azimuth in [45, 315]:
            # Slight perspective transformation (placeholder)
            resized = img.resize((int(img.width * 0.9), img.height))
            view_img = Image.new('RGB', img.size, (128, 128, 128))
            x_offset = (img.width - int(img.width * 0.9)) // 2
            view_img.paste(resized, (x_offset, 0))
        else:
            # Other angles - add border/crop as placeholder
            view_img = img.crop((img.width // 10, 0, 
                               img.width - img.width // 10, img.height))
            view_img = view_img.resize(img.size)
        
        # Save view
        view_filename = f"view_{idx:02d}_{azimuth}az_{elevation}el.png"
        view_path = multiview_path / view_filename
        view_img.save(view_path)
        
        logger.info(f"  [{idx+1}/{len(camera_positions)}] {description} -> {view_filename}")
    
    logger.info(f"\n✓ Saved {len(camera_positions)} placeholder views to: {multiview_path}")
    logger.info("\n" + "=" * 60)
    logger.info("NOTE: These are PLACEHOLDER images!")
    logger.info("=" * 60)
    logger.info("\nTo get real multiview images:")
    logger.info("1. Download Qwen-Image-Edit-2511 (~10GB)")
    logger.info("2. Download Multi-Angles LoRA weights")
    logger.info("3. Run full pipeline")
    logger.info("\nThe full model will generate AI-powered consistent")
    logger.info("multi-angle views from your single input image.")
    
    return multiview_path

def test_trellis_availability():
    """Check if TRELLIS is available"""
    logger.info("\n" + "=" * 60)
    logger.info("Checking 3D Reconstruction Options")
    logger.info("=" * 60)
    
    # Check for TRELLIS
    try:
        import trellis
        logger.info("\n✓ TRELLIS is installed!")
        return True
    except ImportError:
        logger.info("\n✗ TRELLIS not installed")
        logger.info("\nTo install TRELLIS:")
        logger.info("  pip install git+https://github.com/microsoft/TRELLIS.git")
        logger.info("\nAlternatives:")
        logger.info("  - Hunyuan3D-2 (Tencent)")
        logger.info("  - TripoSG (VAST-AI)")
        return False

def create_placeholder_mesh():
    """Create a simple placeholder mesh"""
    logger.info("\n" + "=" * 60)
    logger.info("Creating Placeholder 3D Mesh")
    logger.info("=" * 60)
    
    try:
        import trimesh
        
        # Create a simple sphere as placeholder
        sphere = trimesh.creation.icosphere(subdivisions=3, radius=1.0)
        
        # Save mesh
        output_path = Path("./output")
        output_path.mkdir(exist_ok=True)
        mesh_file = output_path / "placeholder_mesh.ply"
        sphere.export(mesh_file)
        
        logger.info(f"\n✓ Created placeholder mesh:")
        logger.info(f"  - Vertices: {len(sphere.vertices)}")
        logger.info(f"  - Faces: {len(sphere.faces)}")
        logger.info(f"  - Saved to: {mesh_file}")
        logger.info("\nNOTE: This is a PLACEHOLDER sphere mesh!")
        logger.info("Real pipeline will create face-accurate 3D models.")
        
        return mesh_file
        
    except Exception as e:
        logger.error(f"Failed to create mesh: {e}")
        return None

if __name__ == "__main__":
    print("\n" + "=" * 60)
    print("Face to 3D - Simplified Test")
    print("=" * 60)
    print("\nThis is a lightweight test without downloading large models.")
    print("It creates placeholder outputs to demonstrate the pipeline.\n")
    
    # Test 1: Create placeholder multiview images
    multiview_path = create_simple_multiview_demo("test_face.jpg")
    
    # Test 2: Check reconstruction options
    trellis_available = test_trellis_availability()
    
    # Test 3: Create placeholder mesh
    mesh_file = create_placeholder_mesh()
    
    # Summary
    print("\n" + "=" * 60)
    print("Test Summary")
    print("=" * 60)
    print(f"\n[OK] Multiview images: {multiview_path} (8 placeholders)")
    print(f"[OK] 3D mesh: {mesh_file if mesh_file else 'Not created'}")
    print(f"     TRELLIS: {'Available' if trellis_available else 'Not installed'}")
    
    print("\n" + "=" * 60)
    print("Next Steps to Use Real AI Models")
    print("=" * 60)
    print("\n1. For Multiview Generation (Stage 1):")
    print("   - Qwen models are ~10GB")
    print("   - First run downloads models (takes time)")
    print("   - Run: python __init__.py --input test_face.jpg --views 8")
    
    print("\n2. For 3D Reconstruction (Stage 2):")
    print("   - Install TRELLIS:")
    print("     pip install git+https://github.com/microsoft/TRELLIS.git")
    print("   - Or use Hunyuan3D-2 or TripoSG")
    
    print("\n3. Current Status:")
    print("   [OK] Dependencies installed")
    print("   [OK] Test image ready")
    print("   [OK] Pipeline code ready")
    print("   [WARN] Large models not downloaded (would take 10+ minutes)")
    
    print("\n" + "=" * 60)
